+++ src/main/java/frc/robot/subsystems/drive/FastSwerve/ModuleIO.java
@@ -1,7 +1,7 @@
 package frc.robot.subsystems.drive.FastSwerve;
 
 import edu.wpi.first.math.geometry.Rotation2d;
-import frc.robot.utils.selfCheck.drive.SelfChecking;
+import frc.robot.utils.selfCheck.SelfChecking;
 
 import java.util.ArrayList;
 import java.util.List;
