+++ src/main/java/frc/robot/subsystems/drive/FastSwerve/Swerve.java
@@ -30,7 +30,7 @@ import frc.robot.utils.drive.DriveConstants;
 import frc.robot.utils.drive.LocalADStarAK;
 import frc.robot.utils.drive.Sensors.GyroIO;
 import frc.robot.utils.drive.Sensors.GyroIOInputsAutoLogged;
-import frc.robot.utils.selfCheck.drive.SelfChecking;
+import frc.robot.utils.selfCheck.SelfChecking;
 
 import java.util.*;
 import java.util.concurrent.ArrayBlockingQueue;
