+++ src/main/java/frc/robot/subsystems/drive/Mecanum/MecanumIO.java
@@ -4,7 +4,7 @@
 package frc.robot.subsystems.drive.Mecanum;
 
 import edu.wpi.first.math.geometry.Rotation2d;
-import frc.robot.utils.selfCheck.drive.SelfChecking;
+import frc.robot.utils.selfCheck.SelfChecking;
 
 import java.util.ArrayList;
 import java.util.List;
