+++ src/main/java/frc/robot/utils/drive/Sensors/GyroIO.java
@@ -4,7 +4,7 @@
 package frc.robot.utils.drive.Sensors;
 
 import edu.wpi.first.math.geometry.Rotation2d;
-import frc.robot.utils.selfCheck.drive.SelfChecking;
+import frc.robot.utils.selfCheck.SelfChecking;
 
 import java.util.ArrayList;
 import java.util.List;
