+++ src/main/java/frc/robot/utils/selfCheck/drive/SelfCheckingCANCoder.java
@@ -3,6 +3,7 @@ package frc.robot.utils.selfCheck.drive;
 import com.ctre.phoenix6.StatusSignal;
 import com.ctre.phoenix6.hardware.CANcoder;
 
+import frc.robot.utils.selfCheck.SelfChecking;
 import frc.robot.utils.selfCheck.SubsystemFault;
 
 import java.util.ArrayList;
