+++ src/main/java/frc/robot/utils/selfCheck/drive/SelfCheckingPWMMotor.java
@@ -1,6 +1,7 @@
 package frc.robot.utils.selfCheck.drive;
 
 import edu.wpi.first.wpilibj.motorcontrol.PWMMotorController;
+import frc.robot.utils.selfCheck.SelfChecking;
 import frc.robot.utils.selfCheck.SubsystemFault;
 
 import java.util.ArrayList;
